+++ crates/core/src/archive/mod.rs
@@ -15,6 +15,7 @@
 pub mod gzip;
 pub mod tar;
 pub mod targz;
+pub mod tarxz;
 pub mod tarzst;
 pub mod xz;
 pub mod zip;